UML class diagram for my program
Pretty self explanatory.  Most of the functions I created can be found in the Math class in java, i just wanted to take a swing at building them to get a better understanding of how the perform their tasks or different operations.  Could have looked at the source code instead of writing a program to do it.